Remove llm class, add stop_sequences to Chat
