+++ libs/community/langchain_community/llms/cohere.py
@@ -71,7 +71,7 @@ def acompletion_with_retry(llm: Cohere, **kwargs: Any) -> Any:
 
 
 @deprecated(
-    since="0.0.30", removal="0.2.0", alternative_import="langchain_cohere.BaseCohere"
+    since="0.0.30", removal="0.2.0"
 )
 class BaseCohere(Serializable):
     """Base class for Cohere models."""
