+++ libs/partners/cohere/langchain_cohere/__init__.py
@@ -5,7 +5,6 @@ from langchain_cohere.rerank import CohereRerank
 
 __all__ = [
     "ChatCohere",
-    "CohereVectorStore",
     "CohereEmbeddings",
     "CohereRagRetriever",
     "CohereRerank",
